Extract logo into new component
In playing around with the layout, I found myself moving the logo from
component to component. When it is stored as a constant in the component
it is rendered in, it becomes arduous to move it around. As a component,
it is much simpler to move around. Also, it is now reusable in multiple
components, though I doubt that will come in handy any time soon.